refactor(api): split OfferingStatsApi out of providers.rs (#444)
Extract the 4 per-offering statistics handlers (get_provider_offering_stats,
_stats_history, _conversion_stats, _satisfaction_stats) from the monolithic
ProvidersApi in providers.rs into a new OfferingStatsApi in
api/src/openapi/offering_stats.rs.

This is the cleanest cohesive cluster in providers.rs: a tight read-only
analytics group that returns DB-layer types
(crate::database::users::{OfferingStats, OfferingStatsWeek},
crate::database::stats::{OfferingConversionStats, OfferingSatisfactionStats})
— no local request/response types to move, no private-helper entanglement
(default_weeks is a shared pub fn in common.rs; decode_pubkey /
check_authorization are shared). The handlers' serialization tests stay in
providers.rs (they test the DB-layer types, not the handlers).

providers.rs: 4281 → 4090 (−191). Registered OfferingStatsApi in the
create_combined_api tuple; dropped the now-unused default_weeks import from
providers.rs (its sole user moved).

Verified byte-identical via openapi::spec_snapshot — canonical hash
unchanged (de652956…), paths 187 / schemas 327 unchanged, test GREEN.
clippy -p api --tests --all-targets: 0 warnings. offering-stats
serialization tests + spec_snapshot: 6/6 green.

Remaining >2000-line source files under api/src/ (future #444 candidates):
openapi/providers.rs 4090, database/offerings.rs 2876,
database/cloud_resources.rs 2445, openapi/contracts.rs 2244,
openapi/accounts.rs 2230.
