+++ api/src/openapi/providers.rs
@@ -1,5 +1,5 @@
 use super::common::{
-    check_authorization, decode_hex_path, decode_pubkey, default_limit, default_weeks,
+    check_authorization, decode_hex_path, decode_pubkey, default_limit,
     AddAccountContactRequest,
     ApiResponse, ApiTags, AutoAcceptRequest, AutoAcceptResponse,
     BulkUpdatePricesRequest, BulkUpdateStatusRequest,
@@ -3173,197 +3173,6 @@ impl ProvidersApi {
             }),
             error: None,
         })
-    }
-
-    /// Get per-offering contract statistics for a provider
-    ///
-    /// Returns aggregated contract counts and revenue broken down by offering.
-    /// Requires provider authentication — only the provider can access their own stats.
-    #[oai(
-        path = "/providers/:pubkey/offering-stats",
-        method = "get",
-        tag = "ApiTags::Providers"
-    )]
-    async fn get_provider_offering_stats(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-        pubkey: Path<String>,
-    ) -> Json<ApiResponse<Vec<crate::database::users::OfferingStats>>> {
-        let provider_pubkey = match decode_pubkey(&pubkey.0) {
-            Ok(pk) => pk,
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e),
-                });
-            }
-        };
-
-        if let Err(e) = check_authorization(&provider_pubkey, &auth) {
-            return Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e),
-            });
-        }
-
-        match db.get_offering_stats(&provider_pubkey).await {
-            Ok(stats) => Json(ApiResponse {
-                success: true,
-                data: Some(stats),
-                error: None,
-            }),
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(format!("Failed to get offering stats: {e:#}")),
-            }),
-        }
-    }
-
-    /// Get weekly offering stats history for a provider
-    ///
-    /// Returns per-offering weekly contract counts and revenue for the last N weeks.
-    /// Requires provider authentication — only the provider can access their own stats.
-    #[oai(
-        path = "/providers/:pubkey/offering-stats-history",
-        method = "get",
-        tag = "ApiTags::Providers"
-    )]
-    async fn get_provider_offering_stats_history(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-        pubkey: Path<String>,
-        #[oai(default = "default_weeks")] weeks: poem_openapi::param::Query<i32>,
-    ) -> Json<ApiResponse<Vec<crate::database::users::OfferingStatsWeek>>> {
-        let provider_pubkey = match decode_pubkey(&pubkey.0) {
-            Ok(pk) => pk,
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e),
-                });
-            }
-        };
-
-        if let Err(e) = check_authorization(&provider_pubkey, &auth) {
-            return Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e),
-            });
-        }
-
-        let weeks = weeks.0.clamp(1, 52);
-        match db.get_offering_stats_history(&provider_pubkey, weeks).await {
-            Ok(rows) => Json(ApiResponse {
-                success: true,
-                data: Some(rows),
-                error: None,
-            }),
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(format!("Failed to get offering stats history: {e:#}")),
-            }),
-        }
-    }
-
-    /// Get per-offering conversion stats for a provider
-    ///
-    /// Returns views vs rentals breakdown per offering for the last 7 and 30 days.
-    /// Requires provider authentication — only the provider can access their own stats.
-    #[oai(
-        path = "/providers/:pubkey/offering-conversion-stats",
-        method = "get",
-        tag = "ApiTags::Providers"
-    )]
-    async fn get_provider_offering_conversion_stats(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-        pubkey: Path<String>,
-    ) -> Json<ApiResponse<Vec<crate::database::stats::OfferingConversionStats>>> {
-        let provider_pubkey = match decode_pubkey(&pubkey.0) {
-            Ok(pk) => pk,
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e),
-                });
-            }
-        };
-
-        if let Err(e) = check_authorization(&provider_pubkey, &auth) {
-            return Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e),
-            });
-        }
-
-        match db.get_offering_conversion_stats(&provider_pubkey).await {
-            Ok(stats) => Json(ApiResponse {
-                success: true,
-                data: Some(stats),
-                error: None,
-            }),
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(format!("Failed to get offering conversion stats: {e:#}")),
-            }),
-        }
-    }
-
-    /// Get per-offering tenant satisfaction stats for the authenticated provider
-    #[oai(
-        path = "/providers/:pubkey/offering-satisfaction-stats",
-        method = "get",
-        tag = "ApiTags::Providers"
-    )]
-    async fn get_provider_offering_satisfaction_stats(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-        pubkey: Path<String>,
-    ) -> Json<ApiResponse<Vec<crate::database::stats::OfferingSatisfactionStats>>> {
-        let provider_pubkey = match decode_pubkey(&pubkey.0) {
-            Ok(pk) => pk,
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e),
-                });
-            }
-        };
-
-        if let Err(e) = check_authorization(&provider_pubkey, &auth) {
-            return Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e),
-            });
-        }
-
-        match db.get_offering_satisfaction_stats(&provider_pubkey).await {
-            Ok(stats) => Json(ApiResponse {
-                success: true,
-                data: Some(stats),
-                error: None,
-            }),
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(format!("Failed to get offering satisfaction stats: {e:#}")),
-            }),
-        }
     }
 
     /// List auto-accept rules for the authenticated provider
